+++ sdk/storage/azure-storage-files-shares/test/share_client_test.cpp
@@ -345,7 +345,8 @@ namespace Azure { namespace Storage { namespace Test {
 
     {
       std::string directoryName = baseName + RandomString();
-      auto directoryClient = m_shareClient->GetDirectoryClient(directoryName);
+      auto directoryClient
+          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(directoryName);
       EXPECT_NO_THROW(directoryClient.Create());
       auto directoryUrl = directoryClient.GetUri();
       EXPECT_EQ(
@@ -354,7 +355,7 @@ namespace Azure { namespace Storage { namespace Test {
     }
     {
       std::string fileName = baseName + RandomString();
-      auto fileClient = m_shareClient->GetFileClient(fileName);
+      auto fileClient = m_shareClient->GetRootDirectoryClient().GetFileClient(fileName);
       EXPECT_NO_THROW(fileClient.Create(1024));
       auto fileUrl = fileClient.GetUri();
       EXPECT_EQ(fileUrl, m_shareClient->GetUri() + "/" + Storage::Details::UrlEncodePath(fileName));